server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "WhatAreYouDoing? - Action Classification System API Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "google-generativeai>=0.7.1",
    "python-multipart>=0.0.9",
    "SQLAlchemy>=2.0.31",
    "python-dotenv>=1.0.1",
    "Pillow>=10.0.0",
    "tenacity>=8.0.0",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
